Adds organisation API support
Apart from operating a single charge station, the Charge Amps API also
allows controlling (part) of an organization's fleet of charge stations
and users. Since this is most likely an uncommon usecase, I placed it
behind a feature flag "organisations".
This does not add support for CLI flags for accessing the API, but would
be a nice future addition.